automatically backed up by learn

+++ lib/create.sql
@@ -2,7 +2,7 @@ CREATE TABLE projects (
   id INTEGER PRIMARY KEY,
   title TEXT,
   category TEXT,
-  funding_goal INTEGER,
+  funding_goal REAL,
   start_date TEXT,
   end_date TEXT
 );
